requirements.txt
fastapi
uvicorn
celery
redis
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
python-multipart
pypdf
numpy
scikit-learn